Liter gitpod install (#69)
 ### Issues addressed
Our gitpod install is very heavy handed, installing a lot of stuff in the Dockerfile.

This is unwieldy and makes quick dotfile installation difficult.

 ### Summary of changes
Now, we have a fast install and a slow install. Gitpod uses the fast install.

Note that we can't do prebuilds in gitpod because they only save the `/workspace` directory. When this restriction is lifted, we could go back to doing the slow build as a prebuild in gitpod.

 ### Test plan
Started a gitpod, observed that we had our environment.